Post editor no need to-be absoluted element
