+++ web/src/app/admin/admin.module.ts
@@ -5,8 +5,9 @@ import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 
 import { AdminRoutingModule } from './admin-routing.module';
 import { AdminComponent } from './admin.component';
+import { AutoResizeDirective } from './auto-resize.directive';
 import { MyPostsComponent } from './my-posts.component';
-import { PostEditorComponent, AutoResizeDirective } from './post-editor.component';
+import { PostEditorComponent } from './post-editor.component';
 
 import { ApiModule } from '../api/api.module';
 import { GraphQLModule } from '../graphql/graphql.module';
