+++ web/src/app/admin/post-editor.component.scss
@@ -1,12 +1,7 @@
 :host {
   background: #fff;
-  bottom: 0;
   display: block;
-  left: 25.6rem;
-  position: absolute;
-  top: 0;
   width: 100vw;
-  z-index: 100;
 }
 
 .editor {
